refactoring and updating UserControllerTest.java

+++ src/test/java/io/github/samirsales/controller/UserControllerTest.java
@@ -1,45 +1,75 @@
 package io.github.samirsales.controller;
 
-import io.github.samirsales.model.entity.UserEntity;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import io.github.samirsales.model.dto.CredentialDTO;
 import io.github.samirsales.repository.UserRepository;
-import io.github.samirsales.utils.UserEntityGenerator;
-import org.junit.After;
+import org.junit.AfterClass;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.MediaType;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
-import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
+
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @RunWith(SpringRunner.class)
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@SpringBootTest
 @AutoConfigureMockMvc
 public class UserControllerTest {
 
     @Autowired
     @SuppressWarnings("unused")
-    private UserRepository userRepository;
+    private static UserRepository userRepository;
 
     @Autowired
     @SuppressWarnings("unused")
     private MockMvc mockMvc;
 
-
-    @After
-    public void reset(){
-        userRepository.deleteAll();
+    @AfterClass
+    public static void reset(){
+        if(userRepository != null){
+            userRepository.deleteAll();
+        }
     }
 
     @Test
     public void verifyIfGetAllIsForbidden() throws Exception {
-        UserEntity userEntity = UserEntityGenerator.getUserEntityGeneratedById(1L);
-        userRepository.save(userEntity);
-
         ResultActions resultActions = mockMvc.perform(MockMvcRequestBuilders.get("/users/"));
-        resultActions.andExpect(MockMvcResultMatchers.status().isForbidden());
+        resultActions.andExpect(status().isForbidden());
     }
+
+    @Test
+    public void verifyIfGetAllIsOk() throws Exception {
+        String authenticationToken = getAuthenticationToken();
+
+        ResultActions resultActions = mockMvc.perform(MockMvcRequestBuilders.get("/users/")
+            .header(HttpHeaders.AUTHORIZATION, authenticationToken));
+        resultActions.andExpect(status().isOk());
+    }
+
+    private String getAuthenticationToken() throws Exception {
+        ResultActions authenticationResultAction = getAuthenticationResultAction();
+        return authenticationResultAction.andReturn().getResponse().getHeader("Authorization");
+    }
+
+    private ResultActions getAuthenticationResultAction() throws Exception{
+        CredentialDTO credentialDTO = new CredentialDTO("admin", "admin");
+
+        ObjectMapper objectMapper = new ObjectMapper();
+        String credentialDtoStringJson = objectMapper.writeValueAsString(credentialDTO);
+
+        return mockMvc.perform(post("/login")
+            .contentType(MediaType.APPLICATION_JSON)
+            .accept(MediaType.APPLICATION_JSON)
+            .content(credentialDtoStringJson)
+        );
+    }
+
 }
